+++ src/SystemConfig.h
@@ -9,3 +9,11 @@ constexpr byte DEYE_INVERTER_SLAVE_ID = 3;
 
 constexpr byte WAVESHARE_RELAY_CHANNEL_MIN = HouseConfig::Modbus::LIGHT_RELAY_CHANNEL_MIN;
 constexpr byte WAVESHARE_RELAY_CHANNEL_MAX = HouseConfig::Modbus::LIGHT_RELAY_CHANNEL_MAX;
+
+constexpr byte ROLLER_RELAY_SLAVE_ID = HouseConfig::Modbus::ROLLER_RELAY_SLAVE_ID;
+
+constexpr byte ROLLER_RELAY_CHANNEL_MIN = HouseConfig::Modbus::ROLLER_RELAY_CHANNEL_MIN;
+
+constexpr byte ROLLER_RELAY_CHANNEL_MAX = HouseConfig::Modbus::ROLLER_RELAY_CHANNEL_MAX;
+
+constexpr byte ROLLER_RELAY_CHANNEL_COUNT = HouseConfig::Modbus::ROLLER_RELAY_CHANNEL_COUNT;